# AK: Make `Disjoint*::is_empty()` not call size.