This is a raffinement of 49cbd4dcca037336ad5e2e4fcb1e3cc613b46cce.

Previously, the container was scanned to compute the size in the unhappy
path. Now, using `all_of` happy and unhappy path should be fast.