@bgianfo 2021-12-22
>  Now, using `all_of` happy and unhappy path should be fast.

Did you benchmark it to see what the improvement is? 
We tend to request performance numbers if you are optimizing something.