@mhermier 2021-12-22
Do you seriously want me to benchmark `all_of` vs `accumulate` (because this is what the unhappy `size()` path is) ?
I mean the chance that the first chunk/span being not empty is very high and `all_of` will highly probably fails on first iteration... And if the container is empty it becomes a simple iterator/pointer equality. It can be marginally be slower than checking for an integer, but I can't see a world where it would be this visible, unless the compiler does a really poor job at optimizing such trivial code.