@alimpfard 2021-12-22
Performance talk aside for a moment, I like the code better than what we have right now.
Back to performance talk, please understand that the most obvious changes are the most important ones to verify, sure it makes sense that `all_of` would be equivalent or better except in the "is actually empty" case, but we can't be absolutely sure unless we verify that.

That said, this change behaves alright in either extremes, no meaningful change in the worst-case, and a large-ish improvement in the best case.
```
with this change:
    Completed benchmark 'basic_mostly_empty' in 2291ms
    Completed benchmark 'basic_full' in 103ms
    Completed benchmark 'basic_really_empty' in 0ms
without this change:
    Completed benchmark 'basic_mostly_empty' in 2263ms
    Completed benchmark 'basic_full' in 246ms
    Completed benchmark 'basic_really_empty' in 0ms
```

that wasn't so hard, now was it?

<details>
<summary>The benchmarks</summary>

```c++
BENCHMARK_CASE(basic_mostly_empty)
{
    for (size_t i = 0; i < 100; ++i) {
        DisjointChunks<int> chunks;
        for (size_t i = 0; i < 1'000'000; ++i)
            chunks.append({});

        EXPECT(chunks.is_empty());

        chunks.append({1, 2, 3});
        EXPECT(!chunks.is_empty());
    }
}

BENCHMARK_CASE(basic_full)
{
    DisjointChunks<int> chunks;
    for (size_t i = 0; i < 1'000'000; ++i)
        chunks.append({1, 2, 3});

    for (size_t i = 0; i < 100; ++i)
        EXPECT(!chunks.is_empty());
}

BENCHMARK_CASE(basic_really_empty)
{
    DisjointChunks<int> chunks;
    for (size_t i = 0; i < 100; ++i)
        EXPECT(chunks.is_empty());
}
```

</details>