@mhermier 2021-12-22
It is harder than you think, my assumptions have to be verified with a disassembly, but I think with these tests what you are really bench-marking is:
- the memory allocator and the vector implementation for `basic_mostly_empty`
- how smart the compiler is at optimizing loops for `basic_really_empty`.
- both for `basic_full`
Not sure, but maybe it is safer to see if it has impact on the regexp compiler, which seems to be the only user for now...