@alimpfard 2021-12-22
If you think that level of scrutiny is required to verify it, then by all means, please go ahead and do so, because just assuming things and letting them be is not an option - otherwise I can just claim that "we can assume the existing code is more performant, and as testing it is difficult, we won't do so. have a good day.", which is clearly silly.

I believe dodging the two things you mentioned is simple enough, just move the allocation into a separate `noinline` function, and make sure the value can't be considered unchanged through the loop:
```c++
[[gnu::noinline]] void use(auto, auto&) {}
[[gnu::noinline]] auto make_test_vector_mostly_empty()
{
    DisjointChunks<int> chunks;
    for (size_t i = 0; i < 1'000'001; ++i)
        chunks.append({});
    chunks.last_chunk().ensure_capacity(1);
   return chunks;
}

BENCHMARK_CASE(basic_mostly_empty)
{
    auto chunks = make_test_vector_mostly_empty();
    for (size_t i = 0; i < 1000; ++i) {
        use(i, chunks);
    	EXPECT(chunks.is_empty());
	}
    
    chunks.last_chunk().append(1);
    for (size_t i = 0; i < 1000; ++i) {
        use(i, chunks);
    	EXPECT(!chunks.is_empty());
	}
}
```