@mhermier 2021-12-22
I don't see a valid way to properly benchmark this for now without:
- improving the benchmark system (which would require higher precision clocks, introduce resource setup tear-down, and make some statistics about the runs)
- using an external benchmark tool as a temporary band-aid
- or, something I don't see yet...

Is it possible that you rerun your test, with the twist of reserving vectors to expected size in the "Disjoint*" before doing the setup loop? It would be far from perfect but would reduce the impact of the allocations in the test.