@alimpfard 2021-12-22
I don't think the second test I posted does any allocation in the main loop (just once before) - and the allocations pretty much have the same footprint anyway, I don't see how the allocation would affect anything at this point (i.e. outside the loop).
I'll make an issue regarding setup blocks for benchmarks in LibTest.