@mhermier 2021-12-23
Before:
```
  Completed benchmark 'basic_really_empty' in 0ms
  Completed benchmark 'basic_really_empty_large' in 2272ms
  Completed benchmark 'basic_mostly_empty' in 2228ms
  Completed benchmark 'basic_full' in 2367ms
```
After:
```
  Completed benchmark 'basic_really_empty' in 0ms
  Completed benchmark 'basic_really_empty_large' in 2989ms
  Completed benchmark 'basic_mostly_empty' in 3084ms
  Completed benchmark 'basic_full' in 0ms
```

With these we see a good penalty on the extreme cases of ~50%, but there is a dramatic improvement when the case the `DisjointChunks` is full as one would expect to use it.

@alimpfard could you rerun the benchmarks with my change, your numbers are really odd compared to my results, in particular the `basic_full`.